[core] final updates on accounts schema, add go-ch fork with fixes

+++ abi/get.go
@@ -11,9 +11,9 @@ import (
 
 const getMethodsDictKeySz = 19
 
-func MethodNameHash(name string) uint64 {
+func MethodNameHash(name string) uint32 {
 	// https://github.com/ton-blockchain/ton/blob/24dc184a2ea67f9c47042b4104bbb4d82289fac1/crypto/smc-envelope/SmartContract.h#L75
-	return uint64(crc16.Checksum([]byte(name), crc16.MakeTable(crc16.CRC16_XMODEM))) | 0x10000
+	return uint32(crc16.Checksum([]byte(name), crc16.MakeTable(crc16.CRC16_XMODEM))) | 0x10000
 }
 
 func getMethodsDict(code *cell.Cell) (*cell.Dictionary, error) {
